+++ docs/changelog.md
@@ -13,6 +13,7 @@
 ### Fixed
 
 - Don't try to create GitLab comments on unmodified lines - [#1147](https://github.com/cloudflare/pint/pull/1147).
+- [alerts/template](checks/alerts/template.md) check was refactored and will now produce more accurate results.
 
 ## v0.67.0